feat(arc3): improve world model training + code evolution
- World model v2: stop-gradient on targets, grid diff encoder, direct grid reconstruction
- Code evolution v4: 27 executable heuristic codes, better mutation strategies
- Interactive agent v4: lower thresholds (train at 64 buf, code at 0.1 score)
- Fix: info_max_explorer std() on int tensor (cast to float)
- Benchmark focused script for deep single-game exploration

Key findings:
- Transition loss drops from 200+ to 8-14 over 10 episodes
- These are complex sprite-based games (not simple grids)
- Winning requires understanding collision, movement, game rules
- Agent explores but can't discover winning strategies via random actions

+++ soma_mythos_ehra/arc3/info_max_explorer.py
@@ -161,7 +161,7 @@ class InfoMaxExplorer:
             if 0 <= y < grid.shape[0] and 0 <= x < grid.shape[1]:
                 # Score based on local variation
                 patch = grid[max(0,y-1):y+2, max(0,x-1):x+2]
-                score = float(patch.std())
+                score = float(patch.float().std())
                 # Bonus for boundary cells
                 if y > 0 and y < grid.shape[0]-1:
                     if grid[y-1, x] != grid[y+1, x]:
